comp fix

+++ src/test/scala/org/ergoplatform/modifiers/mempool/ErgoTransactionSpec.scala
@@ -573,7 +573,7 @@ class ErgoTransactionSpec extends ErgoPropertyTest with ErgoTestConstants {
     val b = new ErgoBox(1000000000L, Constants.TrueLeaf, Colls.emptyColl,
       Map.empty, ModifierId @@ "c95c2ccf55e03cac6659f71ca4df832d28e2375569cec178dcb17f3e2e5f7742",
       0, 0)
-    val ce = ContextExtension(Map(negId -> IntConstant(0), (-negId) -> IntConstant(1)))
+    val ce = ContextExtension(Map(negId -> IntConstant(0), (-negId).toByte -> IntConstant(1)))
     val input = Input(b.id, ProverResult(Array.emptyByteArray, ce))
 
     val oc = new ErgoBoxCandidate(b.value, b.ergoTree, b.creationHeight)
